Reestructuración de componentes de la página de préstamos
