+++ src/Routes.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import App from "./App";
 import { Home } from "./pages/Home";
-import Notfound from "./components/NavBar/NotFound";
+import Notfound from "./components/Loans/NotFound/NotFound";
 import { createBrowserRouter } from "react-router-dom";
 
 export const routes = [
