+++ src/components/Loans/Loading/Loading.tsx
@@ -1,4 +1,4 @@
-import '../../styles/Loading.css';
+import './Loading.css';
 
 function Loading(){
   return <div className="typewriter">
