+++ src/components/Loans/NotFound/NotFound.tsx
@@ -1,4 +1,4 @@
-import "../../styles/NotFound.css";
+import "./NotFound.css";
 
 function Notfound() {
   return (
