+++ src/components/Loans/check/check.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import "../../styles/NotFound.css";
+import "../NotFound/NotFound.css";
 
 function Notfound() {
   return (
